crud with docker/stocks_products/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
flake8==6.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.2